fix : "스토어 처음 등록시 이미지 안넣은 후 수정하면 에러뜨는 버그 해결"

+++ src/main/java/com/ons/back/application/service/StoreService.java
@@ -92,7 +92,7 @@ public class StoreService {
             store.updateAddress(request.baseAddress(), request.addressDetail());
         }
 
-        if(file != null && !file.isEmpty()) {
+        if(store.getStoreImage() != null && file != null && !file.isEmpty()) {
             storageService.deleteImage(store.getStoreImage());
             store.updateImage(storageService.uploadImage(file));
         }
